+++ src/ck-perf/trace-projections.h
@@ -570,14 +570,14 @@ using namespace PUP;
 
 class toProjectionsFile : public toTextFile {
  protected:
-  virtual void bytes(void *p,int n,size_t itemSize,dataType t);
+  virtual void bytes(void *p,size_t n,size_t itemSize,dataType t);
  public:
   //Begin writing to this file, which should be opened for ascii write.
   toProjectionsFile(FILE *f_) :toTextFile(f_) {}
 };
 class fromProjectionsFile : public fromTextFile {
  protected:
-  virtual void bytes(void *p,int n,size_t itemSize,dataType t);
+  virtual void bytes(void *p,size_t n,size_t itemSize,dataType t);
  public:
   //Begin writing to this file, which should be opened for ascii read.
   fromProjectionsFile(FILE *f_) :fromTextFile(f_) {}
@@ -587,7 +587,7 @@ class fromProjectionsFile : public fromTextFile {
 class toProjectionsGZFile : public PUP::er {
   gzFile f;
  protected:
-  virtual void bytes(void *p,int n,size_t itemSize,dataType t);
+  virtual void bytes(void *p,size_t n,size_t itemSize,dataType t);
  public:
   //Begin writing to this gz file, which should be opened for gz write.
   toProjectionsGZFile(gzFile f_) :er(IS_PACKING), f(f_) {}
